backend/requirements.txt
# ── Core ─────────────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic-settings==2.6.0
pydantic[email]==2.9.2

# ── Supabase ──────────────────────────────────────────────────────────────────
supabase==2.9.1
postgrest==0.17.0

# ── AI SDKs ───────────────────────────────────────────────────────────────────
google-generativeai==0.8.3
groq==0.11.0

# ── HTTP Client ───────────────────────────────────────────────────────────────
httpx==0.27.2
aiohttp==3.10.10
tenacity==9.0.0

# ── Caching ───────────────────────────────────────────────────────────────────
cachetools==5.5.0

# ── Security ──────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# ── Logging ───────────────────────────────────────────────────────────────────
structlog==24.4.0

# ── Date / Timezone ───────────────────────────────────────────────────────────
python-dateutil==2.9.0
pytz==2024.2

# ── Image Processing ──────────────────────────────────────────────────────────
Pillow==11.0.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.3.3
pytest-asyncio==0.24.0
